+++ lib/src/extensions/string.dart
@@ -6,6 +6,8 @@ import 'time.dart';
 final RegExp _emailRegex = RegExp(
     r"^((([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+(\.([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+)*)|((\x22)((((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(([\x01-\x08\x0b\x0c\x0e-\x1f\x7f]|\x21|[\x23-\x5b]|[\x5d-\x7e]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(\\([\x01-\x09\x0b\x0c\x0d-\x7f]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]))))*(((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(\x22)))@((([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))\.)+(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))$");
 final RegExp _percentageRegex = RegExp(r'^[+-]?[0-9]*\.?[0-9]+%$');
+final RegExp _permilleRegex = RegExp(r'^[+-]?[0-9]*\.?[0-9]+‰$');
+final RegExp _permyriadRegex = RegExp(r'^[+-]?[0-9]*\.?[0-9]+‱$');
 final RegExp _urlRegex = RegExp(r'^(https?|ftp):\/\/[^\s/$.?#].[^\s]*$');
 final RegExp _ipv4Regex = RegExp(
     r'^((25[0-5]|2[0-4]\d|1\d{2}|[1-9]?\d)\.){3}(25[0-5]|2[0-4]\d|1\d{2}|[1-9]?\d)$');
@@ -228,14 +230,20 @@ extension StringExtension on String? {
 
   /// Returns double 0.564 from string percentage like "56.4%"
   /// Parse [source] as a double literal and return its value.
-  double? numeric() {
+  num? numeric() {
     if (this == null) {
       return null;
     }
     if (_percentageRegex.hasMatch(this!)) {
-      return double.parse(this!.substring(0, this!.length - 1)) / 100;
+      return num.parse(this!.substring(0, this!.length - 1)) / 100;
     }
-    return double.tryParse(this!);
+    if (_permilleRegex.hasMatch(this!)) {
+      return num.parse(this!.substring(0, this!.length - 1)) / 1000;
+    }
+    if (_permyriadRegex.hasMatch(this!)) {
+      return num.parse(this!.substring(0, this!.length - 1)) / 10000;
+    }
+    return num.tryParse(this!);
   }
 
   /// Returns true if string is a valid JSON string